HEAVY IN DEV PUSH, DONT WORRY ABOUT MERGING MOST OF THIS WILL BE CHANGED
this push is just to make sure no issues come if my computer has file corruption issues or anything like that. This push is not intended to be merged with anything yet as I will definitely be going into this and both heavily editing and culling a lot of content.